Refactoring Toolbar
* Removed Toolbar from Document Content view.
* Added Toolbar to Folder viewer.
* Added Toolbar to PDF viewer.
* Refactored page controls for PDF viewer so they display in Toolbar.

Note: The toolbar is now not displayed in other views. It will be added back to each viewer with affordances for the type of content being displayed.
