+++ aleph/migrate/versions/ea39c04daca4_.py
@@ -1,0 +1,26 @@
+"""empty message
+
+Revision ID: ea39c04daca4
+Revises: 358b9b972b8f
+Create Date: 2018-02-21 14:32:33.828198
+
+"""
+
+# revision identifiers, used by Alembic.
+revision = 'ea39c04daca4'
+down_revision = '358b9b972b8f'
+
+from alembic import op
+import sqlalchemy as sa
+
+
+def upgrade():
+    ### commands auto generated by Alembic - please adjust! ###
+    pass
+    ### end Alembic commands ###
+
+
+def downgrade():
+    ### commands auto generated by Alembic - please adjust! ###
+    pass
+    ### end Alembic commands ###
