+++ ui/src/app/variables.scss
@@ -11,5 +11,6 @@ $aleph-dark-text-color: $pt-dark-text-color;
 $aleph-greyed-text: gray;
 $aleph-content-background: white;
 $aleph-sidebar-background: $light-gray5;
+$aleph-toolbar-background: $light-gray5;
 $aleph-hover-background: $gray5;
 $aleph-border-color: $pt-divider-black;
