+++ ui/src/components/DocumentScreen/DocumentContent.scss
@@ -3,3 +3,11 @@
 .IngestFailed {
     margin-bottom: $aleph-grid-size * 2;
 }
+
+.DualPane .ContentPane.DocumentPane {
+  padding: 0;
+}
+
+.DualPane .ContentPane.DocumentPane .DocumentContent {
+  padding: 20px;
+}
