--- ui/src/components/DocumentScreen/DocumentToolbar.scss
@@ -1,7 +1,0 @@
-@import "src/app/variables.scss";
-
-.document-content-toolbar {
-  margin-bottom: 1em;
-  clear: both;
-  overflow: visible;
-}
