+++ ui/src/components/DocumentScreen/viewers/PdfViewer.jsx
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { Link } from 'react-router-dom';
 import { Document, Page } from 'react-pdf/build/entry.webpack';
 import { findLast, throttle } from 'lodash';
 
 import SectionLoading from 'src/components/common/SectionLoading';
+import DocumentToolbar from 'src/components/common/DocumentToolbar/DocumentToolbar';
 import { parse as parsePdfFragId } from 'src/util/pdfFragId';
 
 import './PdfViewer.css';
@@ -66,22 +66,19 @@ class PdfViewer extends Component {
     if (pageNumber === undefined) pageNumber = 1;
 
     return (
-      <div className="PdfViewer">
-        <div className="pt-button-group pt-fill">
-          <Link to={{ hash: `page=${pageNumber-1}` }}><button className="pt-button pt-icon-arrow-left" disabled={pageNumber <= 1}/></Link>
-          <button className="pt-button pt-fill" disabled={true}>{pageNumber} of {numPages}</button>
-          <Link to={{ hash: `page=${pageNumber+1}` }}><button className="pt-button pt-icon-arrow-right" disabled={pageNumber >= numPages}/></Link>
+      <React.Fragment>
+        <DocumentToolbar document={document} pageNumber={pageNumber} pageTotal={numPages}/>
+        <div className="DocumentContent PdfViewer">
+          <div className="document_pdf"  ref={(ref) => this.pdfElement = ref}>
+            <Document renderAnnotations={true}
+                      file={url}
+                      onLoadSuccess={this.onDocumentLoad.bind(this)}
+                      loading={(<SectionLoading />)}>
+              <Page pageNumber={pageNumber} className="page" width={width} />
+            </Document>
+          </div>
         </div>
-
-        <div className="document_pdf"  ref={(ref) => this.pdfElement = ref}>
-          <Document renderAnnotations={true}
-                    file={url}
-                    onLoadSuccess={this.onDocumentLoad.bind(this)}
-                    loading={(<SectionLoading />)}>
-            <Page pageNumber={pageNumber} className="page" width={width} />
-          </Document>
-        </div>
-      </div>
+      </React.Fragment>
     );
   }
 }
